backend/requirements.txt
Flask==3.1.0
Flask_Cors==5.0.0
newsapi_python==0.2.7
python-dotenv==1.0.1
yfinance==0.2.52
werkzeug==3.1.3
flask-session==0.8.0
DateTime==5.5

